your_image_cabinet/requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.4
django-storages==1.6.5
djangorestframework==3.6.3
google-auth==1.0.2
google-cloud-core==0.26.0
google-cloud-storage==1.3.1
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
gunicorn==19.7.1
h5py==2.7.0
html5lib==0.9999999
idna==2.5
Keras==2.0.6
Markdown==2.6.8
monotonic==1.3
mysqlclient==1.3.10
numpy==1.13.1
olefile==0.44
Pillow==4.2.1
protobuf==3.3.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pytz==2017.2
PyYAML==3.12
requests==2.18.3
rsa==3.4.2
scipy==0.19.1
six==1.10.0
tenacity==4.4.0
tensorflow==1.2.1
Theano==0.9.0
urllib3==1.22
Werkzeug==0.12.2
